feat: update karma type to integer and add quest counters
- Change karmaReward from Decimal to Int in schema
- Enforce integer type for karmaReward in DTO
- Add daily quest counting repository

+++ src/modules/admin/daily-quest/dto/request/create-daily-quest.request.dto.ts
@@ -3,7 +3,7 @@ import { dailyQuestActionValues } from "../../constants/daily-quest-action-optio
 
 export const createDailyQuestRequestSchema = z.object({
   description: z.string().trim().min(1).max(255),
-  karmaReward: z.coerce.number().positive(),
+  karmaReward: z.coerce.number().int().positive(),
   requirement: z.coerce.number().int().positive(),
   action: z.enum(dailyQuestActionValues),
 });
